src/requirements.txt
langgraph==0.0.69
langchain-anthropic
python-dotenv==1.0.1
pillow==10.3.0
fastapi==0.101.1
pydantic==2.4.2
google-cloud-storage==2.10.0
gradio==4.36.1
uvicorn[standard]==0.23.2
python-dotenv==1.0.1
requests==2.32.3